requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langgraph>=0.0.0
streamlit>=1.33.0
playwright>=1.42.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
spacy>=3.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0
reportlab>=4.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=2.0.0
sqlite3-binary>=1.0.0; platform_system == "Linux"
